Refactor lima-and-qemu.pl
The aarch64 version no longer collected the qemu libraries. That's why it was
re-implemented by parsing the `otool -L` output. It eventually turned out that
the Homebrew had switched to Lima 1.0.0, which uses `--vm-type vz` by default
(when available).

The FileMonitor version does not collect kvmvapic.bin and vgabios-virtio.bin.
They may not be needed when the VM is run without a display, but it for now
it seems safer to keep them in the tarball.

The script has also been changed to re-analyze an existing filemonitor.log.
This makes no difference on CI but helps debugging the script.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>